requirements.txt
fastapi
uvicorn
pydantic
langchain>=0.1.16
langchain-google-genai>=0.0.10
google-generativeai>=0.4.1
proto-plus>=1.23.0
protobuf>=4.25.3
python-dotenv
